Inline function which is only called in one place
